+++ src/main/java/com/gft/delivery/model/ItemCompra.java
@@ -1,6 +1,7 @@
 package com.gft.delivery.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.Objects;
 
 import javax.persistence.Entity;
@@ -39,7 +40,7 @@ public class ItemCompra implements Serializable {
 	private int quantity;
 	
 	@NotNull
-	private double price;
+	private BigDecimal price;
 	
 	@ManyToOne
 	@JoinColumn(name = "compra_id")
@@ -69,11 +70,11 @@ public class ItemCompra implements Serializable {
 		this.quantity = quantity;
 	}
 
-	public double getPrice() {
+	public BigDecimal getPrice() {
 		return price;
 	}
 
-	public void setPrice(double price) {
+	public void setPrice(BigDecimal price) {
 		this.price = price;
 	}
